+++ spine-c/spine-c/include/spine/RegionAttachment.h
@@ -39,10 +39,6 @@
 extern "C" {
 #endif
 
-typedef enum {
-	SP_VERTEX_X1 = 0, SP_VERTEX_Y1, SP_VERTEX_X2, SP_VERTEX_Y2, SP_VERTEX_X3, SP_VERTEX_Y3, SP_VERTEX_X4, SP_VERTEX_Y4
-} spVertexIndex;
-
 typedef struct spRegionAttachment {
 	spAttachment super;
 	const char* path;
@@ -64,15 +60,6 @@ void spRegionAttachment_updateOffset (spRegionAttachment* self);
 void spRegionAttachment_computeWorldVertices (spRegionAttachment* self, spBone* bone, float* vertices, int offset, int stride);
 
 #ifdef SPINE_SHORT_NAMES
-typedef spVertexIndex VertexIndex;
-#define VERTEX_X1 SP_VERTEX_X1
-#define VERTEX_Y1 SP_VERTEX_Y1
-#define VERTEX_X2 SP_VERTEX_X2
-#define VERTEX_Y2 SP_VERTEX_Y2
-#define VERTEX_X3 SP_VERTEX_X3
-#define VERTEX_Y3 SP_VERTEX_Y3
-#define VERTEX_X4 SP_VERTEX_X4
-#define VERTEX_Y4 SP_VERTEX_Y4
 typedef spRegionAttachment RegionAttachment;
 #define RegionAttachment_create(...) spRegionAttachment_create(__VA_ARGS__)
 #define RegionAttachment_setUVs(...) spRegionAttachment_setUVs(__VA_ARGS__)