+++ spine-cocos2dx/example/Classes/BatchingExample.cpp
@@ -34,6 +34,8 @@
 USING_NS_CC;
 using namespace spine;
 
+#define NUM_SKELETONS 50
+
 Scene* BatchingExample::scene () {
 	Scene *scene = Scene::create();
 	scene->addChild(BatchingExample::create());
@@ -65,7 +67,7 @@ bool BatchingExample::init () {
 
 	int xMin = _contentSize.width * 0.10f, xMax = _contentSize.width * 0.90f;
 	int yMin = 0, yMax = _contentSize.height * 0.7f;
-	for (int i = 0, j = 0; i < 50; i++) {
+	for (int i = 0, j = 0; i < NUM_SKELETONS; i++) {
 		// Each skeleton node shares the same atlas, skeleton data, and mix times.
 		SkeletonAnimation* skeletonNode = SkeletonAnimation::createWithData(_skeletonData, false);
 		skeletonNode->setAnimationStateData(_stateData);
@@ -76,9 +78,10 @@ bool BatchingExample::init () {
 		
 		// alternative setting two color tint for groups of 10 skeletons
 		// should end up with #skeletons / 10 batches
-		if (j++ < 10)
-			skeletonNode->setTwoColorTint(true);
-		if (j == 20) j = 0;
+		// if (j++ < 10)
+//			skeletonNode->setTwoColorTint(true);
+//		if (j == 20) j = 0;
+		// skeletonNode->setTwoColorTint(true);
 
 		skeletonNode->setPosition(Vec2(
 			RandomHelper::random_int(xMin, xMax),