+++ spine-cocos2dx/src/spine/SkeletonTwoColorBatch.cpp
@@ -168,6 +168,8 @@ SkeletonTwoColorBatch::SkeletonTwoColorBatch () {
 		_commandsPool.push_back(new TwoColorTrianglesCommand());
 	}
 	
+	_indices = spUnsignedShortArray_create(8);
+	
 	reset ();
 	
 	// callback after drawing is finished so we can clear out the batch state
@@ -192,6 +194,8 @@ SkeletonTwoColorBatch::SkeletonTwoColorBatch () {
 
 SkeletonTwoColorBatch::~SkeletonTwoColorBatch () {
 	Director::getInstance()->getEventDispatcher()->removeCustomEventListeners(EVENT_AFTER_DRAW_RESET_POSITION);
+	
+	spUnsignedShortArray_dispose(_indices);
 	
 	for (unsigned int i = 0; i < _commandsPool.size(); i++) {
 		delete _commandsPool[i];
@@ -221,6 +225,36 @@ V3F_C4B_C4B_T2F* SkeletonTwoColorBatch::allocateVertices(uint32_t numVertices) {
 	V3F_C4B_C4B_T2F* vertices = _vertices.data() + _numVertices;
 	_numVertices += numVertices;
 	return vertices;
+}
+	
+	
+void SkeletonTwoColorBatch::deallocateVertices(uint32_t numVertices) {
+	_numVertices -= numVertices;
+}
+
+
+unsigned short* SkeletonTwoColorBatch::allocateIndices(uint32_t numIndices) {
+	if (_indices->capacity - _indices->size < numIndices) {
+		unsigned short* oldData = _indices->items;
+		int oldSize =_indices->size;
+		spUnsignedShortArray_ensureCapacity(_indices, _indices->size + numIndices);
+		unsigned short* newData = _indices->items;
+		for (uint32_t i = 0; i < this->_nextFreeCommand; i++) {
+			TwoColorTrianglesCommand* command = _commandsPool[i];
+			TwoColorTriangles& triangles = (TwoColorTriangles&)command->getTriangles();
+			if (triangles.indices >= oldData && triangles.indices < oldData + oldSize) {
+				triangles.indices = newData + (triangles.indices - oldData);
+			}
+		}
+	}
+	
+	unsigned short* indices = _indices->items + _indices->size;
+	_indices->size += numIndices;
+	return indices;
+}
+
+void SkeletonTwoColorBatch::deallocateIndices(uint32_t numIndices) {
+	_indices->size -= numIndices;
 }
 
 TwoColorTrianglesCommand* SkeletonTwoColorBatch::addCommand(cocos2d::Renderer* renderer, float globalOrder, GLuint textureID, cocos2d::GLProgramState* glProgramState, cocos2d::BlendFunc blendType, const TwoColorTriangles& triangles, const cocos2d::Mat4& mv, uint32_t flags) {
@@ -293,6 +327,7 @@ void SkeletonTwoColorBatch::flush (TwoColorTrianglesCommand* materialCommand) {
 void SkeletonTwoColorBatch::reset() {
 	_nextFreeCommand = 0;
 	_numVertices = 0;
+	_indices->size = 0;
 	_numVerticesBuffer = 0;
 	_numIndicesBuffer = 0;
 	_lastCommand = nullptr;